+++ Changelog.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Fix a bug in fossa-deps.schema.json. It is now valid JSON.
+- Fix a bug in fossa-deps.schema.json. It is now valid JSON. ([#1030](https://github.com/fossas/fossa-cli/pull/1030))
 
 ## 3.3.12